LPD-30649 Move DefinitionInfo component to new DefinitionInfo folder
